refactor: restructure authentication routes and update Discord login flow
