+++ src/routes/(auth)/auth/callback/+server.ts
@@ -1,0 +1,52 @@
+import type { RequestHandler } from "./$types";
+import { redirect, error } from "@sveltejs/kit";
+
+export const GET: RequestHandler = async ({ url, locals: { supabase } }) => {
+  const code = url.searchParams.get("code") as string;
+  const next = url.searchParams.get("next") ?? "/";
+
+  if (!code) throw error(500, "Authorization code is missing.");
+
+  const { data, error: err } = await supabase.auth.exchangeCodeForSession(code);
+  if (err) throw error(500, err.message);
+
+  const { user } = data;
+  if (!user) throw error(500, "User data is missing.");
+
+  // Check if a profile already exists for this user
+  const { data: existingProfile, error: profileFetchError } = await supabase
+    .from("profiles")
+    .select("id")
+    .eq("user_id", user.id)
+    .single();
+
+  if (!profileFetchError && existingProfile) {
+    // Profile exists, just redirect
+    const redirectTo = next.startsWith("/") ? next : `/${next}`;
+    throw redirect(303, redirectTo);
+  }
+
+  // If no profile exists, create one
+  const { data: maxData, error: maxError } = await supabase
+    .from("profiles")
+    .select("id")
+    .order("id", { ascending: false })
+    .limit(1);
+
+  if (maxError) return error(500, { message: maxError.message });
+
+  const newId = maxData?.length ? maxData[0].id + 1 : 1;
+
+  const userId = data.user?.id;
+  const username = data.user.user_metadata.full_name;
+
+  const { error: profileError } = await supabase
+    .from("profiles")
+    .insert({ id: newId, user_id: userId, username, verified: "TRUE" });
+
+  if (profileError) throw error(500, profileError.message);
+
+  // Ensure next always starts with a slash
+  const redirectTo = next.startsWith("/") ? next : `/${next}`;
+  throw redirect(303, redirectTo);
+};
